Rename domain/ to apps/frontend/, create apps/api/ and apps/system/
Three apps, each exporting one router:
- apps/frontend/ — user-facing controllers with before_request hook
- apps/api/ — debug endpoints (DEBUG-only)
- apps/system/ — health check, static files, favicon

server/core.py imports each app's router directly. Removes the
app/routers/ aggregation layer.
