+++ convergence_games/apps/frontend/games/controllers/_detail.py
